Enable the LIBXML_THREAD_ENABLED flag
The LIBXML needs to be compiled with the multi-threads
protection if its resources is shared.
This is the reason why the vanilla version of libxml2
activates this flag by default.

For more safety this patch enables the compilation flag
in xmlversion.h.
Here is the patch upstreamed to google :
https://android-review.googlesource.com/#/c/108830/

Change-Id: I7bf37df8f4ff51cda043aeb2cd6b3b94d1ae455f
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-2784
Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Olivier Fourdan <olivier.fourdan@intel.com>
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: https://android.intel.com:443/276054